refactor: move mock IPC to initializer
